pyproject.toml
[project]
name = "el-fantasy-data-pipeline"
version = "0.1.0"
description = ""
authors = [
    {name = "Milos Marinkovic"}
]
license = {text = "Apache License 2.0"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "boto3 (>=1.37.17,<2.0.0)",
    "numpy (>=2.2.4,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
